Set errors, log in/log out admin

+++ pages/api/products/[id].js
@@ -7,11 +7,14 @@ export default async function handler(req, res) {
     query: { id },
     cookies,
   } = req;
-  // const token = cookies.token
+  const { token } = cookies;
 
   dbConnect();
 
   if (method === "GET") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findById(id);
       res.status(200).json(product);
@@ -35,9 +38,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
-    // if(!token || token !== process.env.token){
-    //   return res.status(401).json("Not authenticated!")
-    // }
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
